# docs/swagger.yaml
swagger: "2.0"
info:
  title: Restfull API
  description: API hệ thống bán lẻ đồ gia dụng
  version: "1.0"
host: "26.107.218.28:5000"
basePath: "/api"
schemes:
  - http
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: "Token phải có dạng 'Bearer <JWT>'"
definitions:
...
  categories:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        parent_id:
          type: integer
  CategoryRequest:
    type: object
    properties:
      name:
        type: string
      parent_id:
        type: integer
  ErrorResponse:
    type: object
    properties:
      message:
        type: string
  SuccessResponse:
    type: object
    properties:
      message:
        type: string
paths:
  /auth/register:
    post:
      tags:
        - auth
      summary: Đăng ký tài khoản
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - email
              - password
            properties:
              username:
                type: string
              email:
                type: string
              password:
                type: string
                format: password
              full_name:
                type: string
              phone_number:
                type: string
              date_of_birth:
                type: string
                format: date
              gender:
                type: string
      responses:
        201:
          description: Đăng ký thành công
        400:
          description: Thiếu thông tin hoặc lỗi đầu vào
        409:
          description: Email hoặc Username đã tồn tại
  /auth/login:
    post:
      tags:
        - auth
      summary: Đăng nhập
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - identity
              - password
            properties:
              identity:
                type: string
              password:
                type: string
                format: password
      responses:
        200:
          description: Đăng nhập thành công, trả về access token
        401:
          description: Sai thông tin đăng nhập
        403:
          description: Tài khoản bị vô hiệu hóa
  /auth/logout:
    post:
      tags:
        - auth
      summary: Đăng xuất
      security:
        - Bearer: []
      responses:
        200:
          description: Đăng xuất thành công
  /auth/forgot-password:
    post:
      tags:
        - auth
      summary: Yêu cầu đặt lại mật khẩu
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - email
            properties:
              email:
                type: string
      responses:
        200:
          description: Đã gửi mã đặt lại mật khẩu
        404:
          description: Email không tồn tại
  /auth/reset-password:
    post:
      tags:
        - auth
      summary: Đặt lại mật khẩu
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - email
              - reset_code
              - new_password
            properties:
              email:
                type: string
              reset_code:
                type: string
              new_password:
                type: string
                format: password
      responses:
        200:
          description: Đặt lại mật khẩu thành công
        422:
          description: Mật khẩu không đủ 6 ký tự 
        403:
          description: Mã không hợp lệ hoặc đã hết hạn
  /auth/change-password:
    put:
      tags:
        - auth
      summary: Thay đổi mật khẩu
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - old_password
              - new_password
            properties:
              old_password:
                type: string
                format: password
              new_password:
                type: string
                format: password
      responses:
        200:
          description: Thay đổi mật khẩu thành công
        400:
          description: Mật khẩu cũ không đúng hoặc lỗi đầu vào
  /auth/request-email-change:
    post:
      tags:
        - auth
      summary: Yêu cầu thay đổi email
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - new_email
            properties:
              new_email:
                type: string
      responses:
        200:
          description: Đã gửi mã xác nhận đến email mới
        400:
          description: Email đã tồn tại
  /auth/change-email:
    put:
      tags:
        - auth
      summary: Xác nhận thay đổi email
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - new_email
              - confirmation_code
            properties:
              new_email:
                type: string
              confirmation_code:
                type: string
      responses:
        422:
          description: Email không đúng format
        409:
          description: Email trùng với email cũ
        200:
          description: Email đã được cập nhật
        400:
          description: Mã xác nhận không hợp lệ hoặc đã hết hạn
  /auth/change-username:
    put:
      tags:
        - auth
      summary: Thay đổi username
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - new_username
            properties:
              new_username:
                type: string
      responses:
        200:
          description: Username đã được cập nhật
        400:
          description: Username đã tồn tại
...
  /category/categories:
    get:
      tags:
        - category
      summary: Lấy danh sách danh mục
      parameters:
        - name: parent_id
          in: query
          required: false
          type: integer
          description: "Lọc theo danh mục cha, nếu không có sẽ lấy tất cả danh mục"
      responses:
        200:
          description: Thành công
          schema:
            type: array
            items:
              $ref: '#/definitions/categories'
  /category/add-category:
    post:
      tags:
        - category
      summary: Thêm danh mục (Chỉ Admin)
      security:
        - Bearer: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/CategoryRequest'
      responses:
        201:
          description: Đã thêm danh mục thành công
          schema:
            $ref: '#/definitions/SuccessResponse'
        400:
          description: Lỗi khi thêm danh mục
          schema:
            $ref: '#/definitions/ErrorResponse'
  /category/update-category/{category_id}:
    put:
      tags:
        - category
      summary: Cập nhật danh mục (Chỉ Admin)
      security:
        - Bearer: []
      parameters:
        - name: category_id
          in: path
          required: false
          type: integer
        - name: body
          in: body
          required: false
          schema:
            $ref: '#/definitions/CategoryRequest'
      responses:
        200:
          description: Đã cập nhật danh mục thành công
          schema:
            $ref: '#/definitions/SuccessResponse'
        404:
          description: Danh mục không tồn tại
          schema:
            $ref: '#/definitions/ErrorResponse'
        304:
          description: không có gì để cập nhật, dữ liệu không được thay đổi
  /category/delete-category/{category_id}:
    delete:
      tags:
        - category
      summary: Xóa danh mục (Chỉ Admin)
      security:
        - Bearer: []
      parameters:
        - name: category_id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: Đã xóa danh mục thành công
          schema:
            $ref: '#/definitions/SuccessResponse'
        404:
          description: Danh mục không tồn tại
          schema:
            $ref: '#/definitions/ErrorResponse'
        400:
          description: Không thể xóa danh mục do có danh mục con hoặc sản phẩm
          schema:
            $ref: '#/definitions/ErrorResponse'
...